[CIR][AArch64] Lower NEON subtraction intrinsics
