+++ clang/test/CodeGen/AArch64/neon-intrinsics.c
@@ -5,198 +5,8 @@
 
 #include <arm_neon.h>
 
-// CHECK-LABEL: define dso_local <8 x i8> @test_vsub_s8(
-// CHECK-SAME: <8 x i8> noundef [[V1:%.*]], <8 x i8> noundef [[V2:%.*]]) #[[ATTR0:[0-9]+]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <8 x i8> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <8 x i8> [[SUB_I]]
-//
-int8x8_t test_vsub_s8(int8x8_t v1, int8x8_t v2) {
-  return vsub_s8(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vsub_s16(
-// CHECK-SAME: <4 x i16> noundef [[V1:%.*]], <4 x i16> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <4 x i16> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <4 x i16> [[SUB_I]]
-//
-int16x4_t test_vsub_s16(int16x4_t v1, int16x4_t v2) {
-  return vsub_s16(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vsub_s32(
-// CHECK-SAME: <2 x i32> noundef [[V1:%.*]], <2 x i32> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <2 x i32> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x i32> [[SUB_I]]
-//
-int32x2_t test_vsub_s32(int32x2_t v1, int32x2_t v2) {
-  return vsub_s32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vsub_s64(
-// CHECK-SAME: <1 x i64> noundef [[V1:%.*]], <1 x i64> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <1 x i64> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <1 x i64> [[SUB_I]]
-//
-int64x1_t test_vsub_s64(int64x1_t v1, int64x1_t v2) {
-  return vsub_s64(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x float> @test_vsub_f32(
-// CHECK-SAME: <2 x float> noundef [[V1:%.*]], <2 x float> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = fsub <2 x float> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x float> [[SUB_I]]
-//
-float32x2_t test_vsub_f32(float32x2_t v1, float32x2_t v2) {
-  return vsub_f32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vsub_u8(
-// CHECK-SAME: <8 x i8> noundef [[V1:%.*]], <8 x i8> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <8 x i8> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <8 x i8> [[SUB_I]]
-//
-uint8x8_t test_vsub_u8(uint8x8_t v1, uint8x8_t v2) {
-  return vsub_u8(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vsub_u16(
-// CHECK-SAME: <4 x i16> noundef [[V1:%.*]], <4 x i16> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <4 x i16> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <4 x i16> [[SUB_I]]
-//
-uint16x4_t test_vsub_u16(uint16x4_t v1, uint16x4_t v2) {
-  return vsub_u16(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vsub_u32(
-// CHECK-SAME: <2 x i32> noundef [[V1:%.*]], <2 x i32> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <2 x i32> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x i32> [[SUB_I]]
-//
-uint32x2_t test_vsub_u32(uint32x2_t v1, uint32x2_t v2) {
-  return vsub_u32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vsub_u64(
-// CHECK-SAME: <1 x i64> noundef [[V1:%.*]], <1 x i64> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <1 x i64> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <1 x i64> [[SUB_I]]
-//
-uint64x1_t test_vsub_u64(uint64x1_t v1, uint64x1_t v2) {
-  return vsub_u64(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vsubq_s8(
-// CHECK-SAME: <16 x i8> noundef [[V1:%.*]], <16 x i8> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <16 x i8> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <16 x i8> [[SUB_I]]
-//
-int8x16_t test_vsubq_s8(int8x16_t v1, int8x16_t v2) {
-  return vsubq_s8(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vsubq_s16(
-// CHECK-SAME: <8 x i16> noundef [[V1:%.*]], <8 x i16> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <8 x i16> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <8 x i16> [[SUB_I]]
-//
-int16x8_t test_vsubq_s16(int16x8_t v1, int16x8_t v2) {
-  return vsubq_s16(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vsubq_s32(
-// CHECK-SAME: <4 x i32> noundef [[V1:%.*]], <4 x i32> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <4 x i32> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <4 x i32> [[SUB_I]]
-//
-int32x4_t test_vsubq_s32(int32x4_t v1, int32x4_t v2) {
-  return vsubq_s32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vsubq_s64(
-// CHECK-SAME: <2 x i64> noundef [[V1:%.*]], <2 x i64> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <2 x i64> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x i64> [[SUB_I]]
-//
-int64x2_t test_vsubq_s64(int64x2_t v1, int64x2_t v2) {
-  return vsubq_s64(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <4 x float> @test_vsubq_f32(
-// CHECK-SAME: <4 x float> noundef [[V1:%.*]], <4 x float> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = fsub <4 x float> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <4 x float> [[SUB_I]]
-//
-float32x4_t test_vsubq_f32(float32x4_t v1, float32x4_t v2) {
-  return vsubq_f32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x double> @test_vsubq_f64(
-// CHECK-SAME: <2 x double> noundef [[V1:%.*]], <2 x double> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = fsub <2 x double> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x double> [[SUB_I]]
-//
-float64x2_t test_vsubq_f64(float64x2_t v1, float64x2_t v2) {
-  return vsubq_f64(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vsubq_u8(
-// CHECK-SAME: <16 x i8> noundef [[V1:%.*]], <16 x i8> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <16 x i8> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <16 x i8> [[SUB_I]]
-//
-uint8x16_t test_vsubq_u8(uint8x16_t v1, uint8x16_t v2) {
-  return vsubq_u8(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vsubq_u16(
-// CHECK-SAME: <8 x i16> noundef [[V1:%.*]], <8 x i16> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <8 x i16> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <8 x i16> [[SUB_I]]
-//
-uint16x8_t test_vsubq_u16(uint16x8_t v1, uint16x8_t v2) {
-  return vsubq_u16(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vsubq_u32(
-// CHECK-SAME: <4 x i32> noundef [[V1:%.*]], <4 x i32> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <4 x i32> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <4 x i32> [[SUB_I]]
-//
-uint32x4_t test_vsubq_u32(uint32x4_t v1, uint32x4_t v2) {
-  return vsubq_u32(v1, v2);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vsubq_u64(
-// CHECK-SAME: <2 x i64> noundef [[V1:%.*]], <2 x i64> noundef [[V2:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = sub <2 x i64> [[V1]], [[V2]]
-// CHECK-NEXT:    ret <2 x i64> [[SUB_I]]
-//
-uint64x2_t test_vsubq_u64(uint64x2_t v1, uint64x2_t v2) {
-  return vsubq_u64(v1, v2);
-}
-
 // CHECK-LABEL: define dso_local <8 x i8> @test_vmul_s8(
-// CHECK-SAME: <8 x i8> noundef [[V1:%.*]], <8 x i8> noundef [[V2:%.*]]) #[[ATTR0]] {
+// CHECK-SAME: <8 x i8> noundef [[V1:%.*]], <8 x i8> noundef [[V2:%.*]]) #[[ATTR0:[0-9]+]] {
 // CHECK-NEXT:  [[ENTRY:.*:]]
 // CHECK-NEXT:    [[MUL_I:%.*]] = mul <8 x i8> [[V1]], [[V2]]
 // CHECK-NEXT:    ret <8 x i8> [[MUL_I]]
@@ -8048,26 +7858,6 @@ int32x4_t test_vqdmlsl_high_s16(int32x4_t a, int16x8_t b, int16x8_t c) {
 //
 int64x2_t test_vqdmlsl_high_s32(int64x2_t a, int32x4_t b, int32x4_t c) {
   return vqdmlsl_high_s32(a, b, c);
-}
-
-// CHECK-LABEL: define dso_local i64 @test_vsubd_s64(
-// CHECK-SAME: i64 noundef [[A:%.*]], i64 noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[VSUBD_I:%.*]] = sub i64 [[A]], [[B]]
-// CHECK-NEXT:    ret i64 [[VSUBD_I]]
-//
-int64_t test_vsubd_s64(int64_t a, int64_t b) {
-  return vsubd_s64(a, b);
-}
-
-// CHECK-LABEL: define dso_local i64 @test_vsubd_u64(
-// CHECK-SAME: i64 noundef [[A:%.*]], i64 noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[VSUBD_I:%.*]] = sub i64 [[A]], [[B]]
-// CHECK-NEXT:    ret i64 [[VSUBD_I]]
-//
-uint64_t test_vsubd_u64(uint64_t a, uint64_t b) {
-  return vsubd_u64(a, b);
 }
 
 // CHECK-LABEL: define dso_local i8 @test_vqaddb_s8(
@@ -18371,16 +18161,6 @@ float64x1_t test_vfma_f64(float64x1_t a, float64x1_t b, float64x1_t c) {
 //
 float64x1_t test_vfms_f64(float64x1_t a, float64x1_t b, float64x1_t c) {
   return vfms_f64(a, b, c);
-}
-
-// CHECK-LABEL: define dso_local <1 x double> @test_vsub_f64(
-// CHECK-SAME: <1 x double> noundef [[A:%.*]], <1 x double> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[SUB_I:%.*]] = fsub <1 x double> [[A]], [[B]]
-// CHECK-NEXT:    ret <1 x double> [[SUB_I]]
-//
-float64x1_t test_vsub_f64(float64x1_t a, float64x1_t b) {
-  return vsub_f64(a, b);
 }
 
 // CHECK-LABEL: define dso_local <1 x double> @test_vabd_f64(
